Fix possible 32 bit issues. In practice hard to reach but better be safe than sorry.

+++ sbin/newfs_udf/newfs_udf.h
@@ -57,7 +57,7 @@ struct udf_create_context context;
 struct udf_disclayout     layout;
 
 /* prototypes */
-int udf_write_sector(void *sector, uint32_t location);
+int udf_write_sector(void *sector, uint64_t location);
 int udf_update_trackinfo(struct mmc_discinfo *di, struct mmc_trackinfo *ti);
 
 /* tmp */
